+++ src/libromdata/tests/CtrKeyScramblerTest.cpp
@@ -89,15 +89,15 @@ class CtrKeyScramblerTest : public ::testing::TestWithParam<CtrKeyScramblerTest_
 		 * Compare two byte arrays.
 		 * The byte arrays are converted to hexdumps and then
 		 * compared using EXPECT_EQ().
-		 * @param expected Expected data.
-		 * @param actual Actual data.
-		 * @param size Size of both arrays.
-		 * @param data_type Data type.
+		 * @param expected	[in] Expected data.
+		 * @param actual	[in] Actual data.
+		 * @param size		[in] Size of both arrays.
+		 * @param data_type	[in] Data type.
 		 */
 		void CompareByteArrays(
 			const uint8_t *expected,
 			const uint8_t *actual,
-			unsigned int size,
+			size_t size,
 			const char *data_type);
 };
 
@@ -105,20 +105,20 @@ class CtrKeyScramblerTest : public ::testing::TestWithParam<CtrKeyScramblerTest_
  * Compare two byte arrays.
  * The byte arrays are converted to hexdumps and then
  * compared using EXPECT_EQ().
- * @param expected Expected data.
- * @param actual Actual data.
- * @param size Size of both arrays.
- * @param data_type Data type.
+ * @param expected	[in] Expected data.
+ * @param actual	[in] Actual data.
+ * @param size		[in] Size of both arrays.
+ * @param data_type	[in] Data type.
  */
 void CtrKeyScramblerTest::CompareByteArrays(
 	const uint8_t *expected,
 	const uint8_t *actual,
-	unsigned int size,
+	size_t size,
 	const char *data_type)
 {
 	// Output format: (assume ~64 bytes per line)
 	// 0000: 01 23 45 67 89 AB CD EF  01 23 45 67 89 AB CD EF
-	const unsigned int bufSize = ((size / 16) + !!(size % 16)) * 64;
+	const size_t bufSize = ((size / 16) + !!(size % 16)) * 64;
 	char printf_buf[16];
 	string s_expected, s_actual;
 	s_expected.reserve(bufSize);
@@ -126,7 +126,7 @@ void CtrKeyScramblerTest::CompareByteArrays(
 
 	// TODO: Use stringstream instead?
 	const uint8_t *pE = expected, *pA = actual;
-	for (unsigned int i = 0; i < size; i++, pE++, pA++) {
+	for (size_t i = 0; i < size; i++, pE++, pA++) {
 		if (i % 16 == 0) {
 			// New line.
 			if (i > 0) {
@@ -135,7 +135,8 @@ void CtrKeyScramblerTest::CompareByteArrays(
 				s_actual += '\n';
 			}
 
-			snprintf(printf_buf, sizeof(printf_buf), "%04X: ", i);
+			// TODO: Print a 64-bit value.
+			snprintf(printf_buf, sizeof(printf_buf), "%04X: ", static_cast<unsigned int>(i));
 			s_expected += printf_buf;
 			s_actual += printf_buf;
 		}
